--- src/main/java/Repository/ApplicationFormRepository.java
@@ -1,5 +1,0 @@
-package Repository;
-
-public interface ApplicationFormRepository {
-
-}
