--- src/main/java/Repository/UserInformationRepository.java
@@ -1,5 +1,0 @@
-package Repository;
-
-public interface UserInformationRepository {
-
-}
