+++ src/main/java/team18/team18_be/apply/repository/ApplyRepository.java
@@ -1,0 +1,5 @@
+package team18.team18_be.apply.repository;
+
+public interface ApplyRepository {
+
+}
