+++ src/main/java/team18/team18_be/userInformation/repository/UserInformationRepository.java
@@ -1,0 +1,5 @@
+package team18.team18_be.userInformation.repository;
+
+public interface UserInformationRepository {
+
+}
